+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## [0.0.1] - 2022.07.03
+- deleteSave function added
 - loadGame function added
 - saveGame function added